The package bigtop-tomcat was included because of distributions like CentOS 5.x didn't shipped with recent versions of Tomcat [1]. Due to Buildoop support >= CentOS 6.x and those versions are shipping Tomcat 6, a more recent version of this Servlet Container  and Webserver, this dependency of "in-house Tomcat package" was removed.
Because Oozie, and others need system-wide Tomcat server, those
packages need depends on the standard distribution Tomcat package.
The problem here is when we install the standard package "yum install tomcat6"
this installation is plenty of dependencies of other packages, and
one of them is the OpenJDK dependency, we need to avoid this dependency,
because we are only supporting Oracle Sun JDKs. At this point we need
to include a "in-house Tomcat package" as Bigtop project does it.

[1] https://issues.apache.org/jira/browse/BIGTOP-417

This package is related with the issue [buildoop-github-#33]