This package is an example for inyect Avro files to Kafka using the "Avro Server Schema Repository" used by Camus kafka consummer. This version is downloaded from [1]. And is packaged from the release tag "release-0.8.1". The nomenclature version rationale is the following:
The producer is linked with the Kafka version 0.8, and this
version of producer is the "version 1". So the tag is named
"release-0.8.1". The last number will change with new version
of this producer, and the two first numbers only will be changed
with the mayor version of Kafka.

[1] https://github.com/buildoop/AvroRepoKafkaProducerTest

This commit fix #78